[FIX] 마이페이지 모바일 수정]
